requirements.txt
tensorflow==2.8.4
fastapi
uvicorn
numpy
matplotlib
pillow
python-multipart